Move `keyFromAuthData` to `common-crypto/key-passphrase.ts`
